// src/API.ts
export enum TableName {
  User = "User",
  Profile = "Profile",
  MediaFile = "MediaFile",
  Post = "Post",
  Comment = "Comment",
}
// ...
export type User = {
  __typename: "User",
  id: string,
  type: TableName,
  name: string,
  email: string,
  userOwnerId?: string | null,
  userOwnerName?: string | null,
  accountConfirmed?: boolean | null,
  createdAt?: string | null,
  profileCreated?: boolean | null,
  owner?: string | null,
  updatedAt: string,
  profile?: Profile | null,
  mediaFiles?: ModelMediaFileConnection | null,
  posts?: ModelPostConnection | null,
  comments?: ModelCommentConnection | null,
};

export type Profile = {
  __typename: "Profile",
  id: string,
  type: TableName,
  userId: string,
  classRoom?: string | null,
  userEmail: string,
  classRoomYear?: string | null,
  schoolLeavingYear?: string | null,
  year?: string | null,
  role: Role,
  roleInteger: number,
  profileOwnerId?: string | null,
  profileOwnerName?: string | null,
  owner?: string | null,
  createdAt: string,
  updatedAt: string,
  user?: User | null,
};

export enum Role {
  TeachingStaff = "TeachingStaff",
  Student = "Student",
  NonTeachingStaff = "NonTeachingStaff",
  Alumni = "Alumni",
  ExTeachingStaff = "ExTeachingStaff",
  ExNonTeachingStaff = "ExNonTeachingStaff",
  Headmaster = "Headmaster",
  AsstHeadMaster = "AsstHeadMaster",
  AlumniAndTeachingStaff = "AlumniAndTeachingStaff",
  AlumniAndNonTeachingStaff = "AlumniAndNonTeachingStaff",
  AlumniAndExTeachingStaff = "AlumniAndExTeachingStaff",
  AlumniAndExNonTeachingStaff = "AlumniAndExNonTeachingStaff",
  Admin = "Admin",
}


export type ModelMediaFileConnection = {
  __typename: "ModelMediaFileConnection",
  items?:  Array<MediaFile | null > | null,
  nextToken?: string | null,
};

export type MediaFile = {
  __typename: "MediaFile",
  id: string,
  type: TableName,
  userId?: string | null,
  userEmail: string,
  file:  Array<S3Object | null >,
  createdAt: string,
  owner?: string | null,
  updatedAt: string,
  user?: User | null,
};

export type S3Object = {
  __typename: "S3Object",
  bucket: string,
  region: string,
  key: string,
};

export type ModelPostConnection = {
  __typename: "ModelPostConnection",
  items?:  Array<Post | null > | null,
  nextToken?: string | null,
};

export type Post = {
  __typename: "Post",
  type: TableName,
  id: string,
  userId?: string | null,
  userEmail: string,
  createdAt: string,
  updatedAt: string,
  title?: string | null,
  content?: string | null,
  files?:  Array<S3Object | null > | null,
  likes?: Array< string | null > | null,
  unlikes?: Array< string | null > | null,
  likeCount?: string | null,
  unlikeCount?: string | null,
  owner?: string | null,
  user?: User | null,
  comments?: ModelCommentConnection | null,
};

export type ModelCommentConnection = {
  __typename: "ModelCommentConnection",
  items?:  Array<Comment | null > | null,
  nextToken?: string | null,
};

export type Comment = {
  __typename: "Comment",
  type: TableName,
  id: string,
  postId: string,
  userId?: string | null,
  userEmail: string,
  content?: string | null,
  createdAt: string,
  updatedAt: string,
  likes?: Array< string | null > | null,
  unlikes?: Array< string | null > | null,
  likeCount?: string | null,
  unlikeCount?: string | null,
  owner?: string | null,
  user?: User | null,
  post?: Post | null,
};